+++ contract-deployment/package.json
@@ -1,15 +1,13 @@
 {
-  "name": "@aztec-fpc/attestation",
+  "name": "@aztec-fpc/contract-deployment",
   "version": "0.1.0",
+  "description": "Bundleable FPC contract deployment CLI for devnet",
   "type": "module",
-  "description": "Quote-signing attestation service for the MultiAssetFPC",
   "main": "dist/index.js",
   "scripts": {
     "build": "bun build src/index.ts --outdir=dist --target=bun --external @aztec/bb.js --external @aztec/noir-noirc_abi --external @aztec/noir-acvm_js --external pino",
     "typecheck": "tsc --noEmit",
-    "test": "tsx --test test/*.test.ts",
-    "start": "bun run dist/index.js",
-    "dev": "tsx src/index.ts"
+    "start": "bun run dist/index.js"
   },
   "dependencies": {
     "@aztec/accounts": "4.0.0-devnet.2-patch.3",
@@ -18,16 +16,10 @@
     "@aztec/foundation": "4.0.0-devnet.2-patch.3",
     "@aztec/noir-noirc_abi": "4.0.0-devnet.2-patch.3",
     "@aztec/noir-acvm_js": "4.0.0-devnet.2-patch.3",
-    "@aztec/pxe": "4.0.0-devnet.2-patch.3",
     "@aztec/stdlib": "4.0.0-devnet.2-patch.3",
     "@aztec/wallets": "4.0.0-devnet.2-patch.3",
-    "fastify": "^4.28.0",
     "pino": "^10.3.1",
-    "pino-pretty": "^13.1.3",
-    "viem": "^2.18.0",
-    "yaml": "^2.4.5",
-    "zod": "^3.23.8",
-    "@aztec-fpc/contract-deployment": "workspace:*"
+    "pino-pretty": "^13.1.3"
   },
   "engines": {
     "bun": "1.3.9"
